refactor: 계산기 클래스를 별도로 분리
